+++ apps/admin-ui-dashboard/src/views/app-view.tsx
@@ -5,7 +5,6 @@
  */
 import { Container } from '@zextras/ui-components';
 import { FC, Suspense } from 'react';
-import { Route, Routes } from 'react-router';
 
 import { Breadcrumb } from './breadcrumb/breadcrumb';
 import Dashboard from './dashboard/dashboard-view';
@@ -18,15 +17,12 @@ const DashboardComponent = () => (
   </Container>
 );
 
-const AppView: FC = () => {
+export const AppView: FC = () => {
   return (
     <Container height={'fit'}>
       <Breadcrumb />
-      <Routes>
-        <Route path={'/*'} element={<DashboardComponent />} />
-      </Routes>
+      <DashboardComponent />
     </Container>
   );
 };
 
-export default AppView;