+++ apps/admin-ui-domains/src/views/domain/manange/accounts/edit-account/edit-account-general-section.tsx
@@ -19,7 +19,7 @@ import { Trans, useTranslation } from 'react-i18next';
 
 import type { Attribute, objectType } from '../../../../../../types';
 import type { SoapEntity } from '../../../../../../types/service';
-import { ADMINISTRATION, DEFAULT, MAX_DOMAIN_DISPLAY, TRUE } from '../../../../../constants';
+import { ADMINISTRATION, DEFAULT, MAX_DOMAIN_DISPLAY } from '../../../../../constants';
 import { useSelectedDomain } from '../../../../../hooks/use-selected-domain';
 import { endSession } from '../../../../../services/end-session';
 import { getDelegateAuthRequest } from '../../../../../services/get-delegate-auth-request';
@@ -73,8 +73,6 @@ export const EditAccountGeneralSection: FC<{
     setAllUserSessionList,
     userSessionList,
     setUserSessionList,
-    defaultCOS,
-    setDefaultCOS,
     allowedDeletePassword,
   } = useContext(AccountContext);
   const { data: domain } = useSelectedDomain();
@@ -101,6 +99,8 @@ export const EditAccountGeneralSection: FC<{
   const [isRequestInProgress, setIsRequestInProgress] = useState<boolean>(false);
   const isAdvanced = useIsAdvanced();
   const [defaultCosId, setDefaultCosId] = useState('');
+  const [defaultCOS, setDefaultCOS] = useState<boolean>(false);
+  const [cosDefaultStateSet, setCosDefaultStateSet] = useState<boolean>(false);
 
   const sessionTableHeader: any[] = useMemo(() => {
     const accountsLabel = t('label.accounts', 'Accounts');
@@ -236,11 +236,12 @@ export const EditAccountGeneralSection: FC<{
     if (accountDetail?.zimbraCOSId) {
       const cosId = cosItems.find((item: any) => item.label === DEFAULT);
       setDefaultCosId(cosId?.value);
-      if (accountDetail?.zimbraCOSId === cosId?.value) {
-        setDefaultCOS(TRUE);
+      if (accountDetail?.zimbraCOSId === cosId?.value && !cosDefaultStateSet) {
+        setDefaultCOS(true);
+        setCosDefaultStateSet(true);
       }
     }
-  }, [accountDetail, cosItems, setDefaultCOS]);
+  }, [accountDetail, cosItems, defaultCOS, setDefaultCOS, cosDefaultStateSet]);
 
   const selection = useMemo(
     () => cosItems.find((item: any) => item.value === accountDetail?.zimbraCOSId),