+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## [0.15.2](https://github.com/zextras/carbonio-admin-console-ui/compare/v0.15.1...v0.15.2) (2026-07-20)
+
+### Bug Fixes
+
+* fixed default cos enable switch issue ([#1296](https://github.com/zextras/carbonio-admin-console-ui/issues/1296)) ([815d48e](https://github.com/zextras/carbonio-admin-console-ui/commit/815d48e5250491caa51ff87ce145e310c93b67af))
+
+## [0.15.1](https://github.com/zextras/carbonio-admin-console-ui/compare/v0.15.0...v0.15.1) (2026-07-20)
+
+### Bug Fixes
+
+* fixed event bubbling during cos delete ([#1299](https://github.com/zextras/carbonio-admin-console-ui/issues/1299)) ([0705547](https://github.com/zextras/carbonio-admin-console-ui/commit/0705547acdae1a3ec5a1ae93f161596a72facebd))
+
+## [0.15.0](https://github.com/zextras/carbonio-admin-console-ui/compare/v0.14.7...v0.15.0) (2026-07-16)
+
+### Features
+
+* **routing:** browser refresh keeps you in the same view ([ca337f2](https://github.com/zextras/carbonio-admin-console-ui/commit/ca337f2ffa8bbd0653456948bae9eacfeea7d400))
+
 ## [0.14.7](https://github.com/zextras/carbonio-admin-console-ui/compare/v0.14.6...v0.14.7) (2026-07-08)
 
 ### Other changes